+++ examples/gltf/src/main.rs
@@ -1,5 +1,3 @@
-use glam::UVec2;
-
 fn load_gltf(
     renderer: &rend3::Renderer,
     path: &'static str,
@@ -74,7 +72,7 @@ fn main() {
         &surface,
         &iad.device,
         format,
-        UVec2::new(window_size.width, window_size.height),
+        glam::UVec2::new(window_size.width, window_size.height),
         rend3::types::PresentMode::Mailbox,
     );
 
@@ -85,7 +83,7 @@ fn main() {
     let mut routine = rend3_pbr::PbrRenderRoutine::new(
         &renderer,
         rend3_pbr::RenderTextureOptions {
-            resolution: UVec2::new(window_size.width, window_size.height),
+            resolution: glam::UVec2::new(window_size.width, window_size.height),
             samples: rend3_pbr::SampleCount::Four,
         },
         format,
@@ -105,15 +103,14 @@ fn main() {
     // We need to keep the object alive.
     let _object_handle = renderer.add_object(object);
 
+    let view_location = glam::Vec3::new(3.0, 3.0, -5.0);
+    let view = glam::Mat4::from_euler(glam::EulerRot::XYZ, -0.49, 0.43, 0.0);
+    let view = view * glam::Mat4::from_translation(-view_location);
+
     // Set camera's location
     renderer.set_camera_data(rend3::types::Camera {
-        projection: rend3::types::CameraProjection::Projection {
-            vfov: 60.0,
-            near: 0.1,
-            pitch: 0.43,
-            yaw: -0.49,
-        },
-        location: glam::Vec3A::new(3.0, 3.0, -5.0),
+        projection: rend3::types::CameraProjection::Projection { vfov: 60.0, near: 0.1 },
+        view,
     });
 
     // Create a single directional light
@@ -140,13 +137,13 @@ fn main() {
             event: winit::event::WindowEvent::Resized(size),
             ..
         } => {
-            let size = UVec2::new(size.width, size.height);
+            let size = glam::UVec2::new(size.width, size.height);
             // Reconfigure the surface for the new size.
             rend3::configure_surface(
                 &surface,
                 &renderer.device,
                 format,
-                UVec2::new(size.x, size.y),
+                glam::UVec2::new(size.x, size.y),
                 rend3::types::PresentMode::Mailbox,
             );
             // Tell the renderer about the new aspect ratio.