+++ rend3-pbr/src/culling/gpu.rs
@@ -291,7 +291,7 @@ impl GpuCuller {
         if let Some(sorting) = args.sort {
             profiling::scope!("Sorting");
 
-            let camera_location = args.camera.get_data().location;
+            let camera_location = args.camera.location().into();
 
             match sorting {
                 Sorting::FrontToBack => {